Clean up shared tree provider code
1. Rename files to match class names
1. Remove some unused imports
