+++ src/CosmosEditorManager.ts
