+++ src/mongo/MongoCommand.ts
@@ -1,0 +1,14 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as vscode from 'vscode';
+
+export interface MongoCommand {
+    range: vscode.Range;
+    text: string;
+    collection?: string;
+    name: string;
+    arguments?: string;
+}
