+++ src/mongo/editors/MongoFindOneResultEditor.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { MongoCollectionTreeItem, MongoDatabaseTreeItem } from "../nodes";
 import { IAzureParentNode, IAzureNode } from "vscode-azureextensionui";
 import { IMongoDocument, MongoDocumentTreeItem } from "../tree/MongoDocumentTreeItem";
-import { ICosmosEditor } from "../../DocumentEditor";
+import { ICosmosEditor } from "../../CosmosEditorManager";
+import { MongoDatabaseTreeItem } from "../tree/MongoDatabaseTreeItem";
+import { MongoCollectionTreeItem } from "../tree/MongoCollectionTreeItem";
 
 export class MongoFindOneResultEditor implements ICosmosEditor<IMongoDocument> {
     private _databaseNode: IAzureParentNode<MongoDatabaseTreeItem>;
